requirements.txt
# Smart Productivity Analytics Platform - Requirements
# ===================================================

# Core Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-spectacular>=0.27.0

# Filtering
django-filter>=23.5

# CORS
django-cors-headers>=4.3.0

# Database (optional - for production)
# psycopg2-binary>=2.9.0  # PostgreSQL

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0

# HTTP requests (for external API integration)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Utility
python-dateutil>=2.8.0
